test folder cleanup
