+++ Tests/VaporTests/Utilities/ByteBuffer+Helpers.swift
@@ -1,0 +1,9 @@
+import Vapor
+
+extension ByteBuffer {
+    init(string: String) {
+        var buffer = ByteBufferAllocator().buffer(capacity: 0)
+        buffer.writeString(string)
+        self = buffer
+    }
+}
